Update the licensing information
The previous page was a bit vague about how the code was licensed under
MIT and content under CC BY (since the distinction between code and
content is unclear).  Clarify the wording to indicate that everything is
available under CC BY (with exceptions) and code and markup are also
under MIT.

Add the programs to the exceptions list, just to further clarify that
programs are covered by their own licenses.

Link to the source on GitHub.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>