Big .htaccess cleanup
Fix the RewriteRules to use correct variable names and to preserve the
query string when rewriting (currently unused, but may be in the
future).

Lots and lots of debugging to get MultiViews working for serving XHTML.
This is an unholy abomination, but it works.  I hope Apache gets a fix
for the qs parameter eventually.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>