Rename the Atom feed
Move the Atom feed into a feeds directory below bits.  This directory
will hold all of the feeds for the blog (if/when tag feeds are added).

Also change the extension from .xml to .atom.  This makes it easier to
serve it with the correct media type on the server and to distinguish it
from any future RSS XML.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>