Increment markdown headers by 2 levels
We have the page title and article title at higher levels, so any
headers in the post should start at h3.

I'm not sure this is an ideal solution, but since I didn't find a way to
increment the header level in the markdown -> HTML generator, I'll stick
with this for the moment.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>