Re-add the Vim Scripts
Even though the scripts have been integrated into the current releases
of Vim, there's no need for the information to vanish into history.
Update the page to make it clear the script is included in current
releases and retain the information for historical reference and for
users still using older versions of Vim.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>