Rework the site style and navigation
The old layout with the sidebar didn't scale to small sizes very well as
the sidebar took up most of the screen space.  Rework the navigation
(the only thing in the sidebar) to display below the site name.  Update
styles to be a bit simpler with a plain single-column layout.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>