Add xml:base to Atom post content
Posts which contain relative links (e.g. to images) won't work correctly
in the Atom feed, since the URLs are resolved relative to the location
of the feed rather than the original page.  Correct this by adding an
xml:base which makes links in the content relative to the original page.

TODO:  Check if there are any popular feed readers which don't support
xml:base.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>