Store content separately from posts
Previously I used a patched version of Jekyll while waiting on a merge
decision from upstream.  Now that the decision has been made and the
changes will not be included (nor similar attempts to store contents
with posts), it's not worth the effort to maintain these changes.
Remove the content-with-post plugin and place the content in the
destination directory manually.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>