Update to Jekyll 2.2.0
Update our configuration, plugins, and invocations to work with the
version of Jekyll currently packaged in Debian Testing.

We have fallen far behind due to the custom patches for post content
that we were holding on to.  Now that those are gone, update to the
latest unpatched version available in Debian.

This also allows us to get rid of the post-extras plugin, which is
superseded by the excerpt_separator configuration option.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>