Disable content negotiation for ErrorDocuments
Enabling content negotiation for ErrorDocuments (by not specifying a
file extension) can cause 406 Not Acceptable errors on top of previous
errors in the case that the client doesn't accept any of the available
ErrorDocument types (e.g. only accepting image/gif when requesting a
.gif URL).  This results in sending a generic, generated error page
about the double-error.

To avoid this, check for application/xhtml+xml in the Accept header and
send that if present, otherwise send text/html.  Although this is not
perfect (e.g. doesn't do preference/quality comparison), it works well
in practice since the only user agents which mention
application/xhtml+xml accept it with equal preference to text/html.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>