Fix serving of .gz files
By default (on Debian) .gz files are interpreted as files of type
application/x-gzip.  Although our AddEncoding directive worked correctly
for most negotiated requests, it would fail when the negotiated type was
not explicitly accepted by the client.  When this occurred, either a 406
error would be generated with incorrect TCN information (if none of the
available types are accepted) or a response of application/x-gzip would
be sent with a gzip encoding (resulting in ugly errors).

Fix this by removing the type for the .gz extension, except when serving
.tar.gz which should be served with this type.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>