Set page permalink style to "none"
Due to the changes in jekyll/jekyll@90865d (Jekyll 3.1.0+) .xhtml pages
are not treated like .html pages which are placed in subdirectories
based on the sitewide permalink setting (from jekyll/jekyll#3538).

To avoid this, set the sitewide permalink setting to "none" and set a
permalink default value for drafts and pages with the desired permalink
for those types.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>