Add more conditions to deflate FilterProvider
Although the deflate filter does not make any changes to requests which
do not need compression applied, mod_filter does change the headers
whenever a FilterProvider matches (which includes removing the Etag
header and setting Accept-Range: none for deflate).  To avoid this, make
the filter conditional on Accept-Encoding and the response encoding.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>