Update brotli command arguments for 1.0
The 1.0 release of the brotli command replaced --input and --output with
a more gzip-like convention of accepting filename arguments and a
-S/--suffix option to determine the resulting filename.  Use it.

Also comment out the permission workaround required for previous
versions, since they would fail to run with the new syntax.  Can
uncomment if the old version needs to be supported in the future.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>