build-multiviews: Split file building script
It's sometimes convenient to be able to (re)build the MultiViews
variants for specific files, rather than every file in a directory.
Move the code for doing this into a script called build-multiview and
call it from build-multiviews as appropriate.

Note: This also avoids the need for a temporary file of compressible
file names.  Nice!

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>